changed cache to use Guava instead of a separate jar
