+++ cas-server-core/src/main/java/org/jasig/cas/util/services/RegisteredServiceJsonSerializer.java
@@ -25,7 +25,6 @@ import org.jasig.cas.services.RegisteredServiceAccessStrategy;
 import org.jasig.cas.services.RegisteredServiceProxyPolicy;
 import org.jasig.cas.util.AbstractJacksonBackedJsonSerializer;
 
-import javax.cache.expiry.Duration;
 import java.net.URL;
 import java.util.Map;
 
@@ -52,26 +51,10 @@ public final class RegisteredServiceJsonSerializer extends AbstractJacksonBacked
         final ObjectMapper mapper = super.initializeObjectMapper();
         mapper.addMixIn(RegisteredServiceProxyPolicy.class, RegisteredServiceProxyPolicyMixin.class);
         mapper.addMixIn(RegisteredServiceAccessStrategy.class, RegisteredServiceAuthorizationStrategyMixin.class);
-        mapper.addMixIn(Duration.class, DurationMixin.class);
+
         return mapper;
     }
 
-    private static class DurationMixin extends Duration {
-
-        private static final long serialVersionUID = 743505593336053306L;
-        @JsonIgnore
-        @Override
-        public boolean isEternal() {
-            return false;
-        }
-
-        @JsonIgnore
-        @Override
-        public boolean isZero() {
-            return false;
-        }
-
-    }
     private static class RegisteredServiceProxyPolicyMixin implements RegisteredServiceProxyPolicy {
 
         private static final long serialVersionUID = 4854597398304437341L;
