+++ cas-server-core/src/test/java/org/jasig/cas/services/JsonServiceRegistryDaoTests.java
@@ -20,9 +20,8 @@ package org.jasig.cas.services;
 
 import com.google.common.collect.Sets;
 import org.apache.commons.io.FileUtils;
-import org.jasig.cas.authentication.principal.cache.CachingPrincipalAttributesRepository;
 import org.jasig.cas.authentication.principal.ShibbolethCompatiblePersistentIdGenerator;
-import org.jasig.cas.authentication.principal.cache.GuavaCachingPrincipalAttributesRepository;
+import org.jasig.cas.authentication.principal.cache.CachingPrincipalAttributesRepository;
 import org.jasig.cas.services.support.RegisteredServiceRegexAttributeFilter;
 import org.jasig.services.persondir.support.StubPersonAttributeDao;
 import org.jasig.services.persondir.support.merger.ReplacingAttributeAdder;
@@ -294,8 +293,8 @@ public class JsonServiceRegistryDaoTests {
         final Map<String, List<Object>> attributes = new HashMap<>();
         attributes.put("values", Arrays.asList(new Object[]{"v1", "v2", "v3"}));
 
-        final GuavaCachingPrincipalAttributesRepository repository =
-                new GuavaCachingPrincipalAttributesRepository(
+        final CachingPrincipalAttributesRepository repository =
+                new CachingPrincipalAttributesRepository(
                         new StubPersonAttributeDao(attributes),
                         TimeUnit.MILLISECONDS, 100);
         repository.setMergingStrategy(new ReplacingAttributeAdder());
